refactor(sim): extract SLOPriorityMap to slo_priority.go
Move SLOPriorityMap type and methods from admission.go to a dedicated
file. Same package, zero import changes. Admission policies unchanged.
